refactor: extract lifecycle flows
